+++ lint-staged.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   'package.json': 'sort-package-json',
-  '*.{ts,tsx}': 'eslint --max-warnings=0 . --fix',
+  '*.{ts,tsx}': 'eslint --no-warn-ignored --max-warnings=0 . --fix',
   '**/*.ts?(x)': () => 'tsc -p tsconfig.json',
 };